+++ src/algorithms/telemetry_decoder/gnuradio_blocks/beidou_b3i_telemetry_decoder_gs.h
@@ -20,6 +20,7 @@
 #include "beidou_dnav_navigation_message.h"
 #include "gnss_block_interface.h"
 #include "gnss_satellite.h"
+#include "nav_message_packet.h"
 #include "tlm_conf.h"
 #include <boost/circular_buffer.hpp>
 #include <gnuradio/block.h>  // for block
@@ -85,6 +86,8 @@ private:
     // Navigation Message variable
     Beidou_Dnav_Navigation_Message d_nav;
 
+    Nav_Message_Packet d_nav_msg_packet;
+
     Gnss_Satellite d_satellite;
 
     std::string d_dump_filename;
@@ -116,6 +119,7 @@ private:
     bool d_dump;
     bool d_dump_mat;
     bool d_remove_dat;
+    bool d_enable_navdata_monitor;
 };